Change directory name to components
